+++ scripts/actions/patch.mjs
@@ -43,8 +43,9 @@ export default async function () {
 }
 function gerVersion() {
    try {
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const source = process.env.BUILD_SOURCE;
-      return `v${packageJson?.version ?? '0.0.1'}${source ? `|${source}` : ''}`;
+      return `v${packageJson?.version ?? '0.0.1'}`; //${source ? `|${source}` : ''}
    } catch {
       return 'v0.0.1';
    }